requirements.txt
pandas==2.1.4
duckdb==0.9.2
pyyaml==6.0.1
python-dotenv==1.0.0